Move Context to a separate file.
